+++ src/svg.rs
@@ -1,7 +1,6 @@
 use bevy::{
     asset::Handle,
-    ecs::component::Component,
-    math::{Mat4, Vec2, Vec3},
+    math::{Mat4, Vec2},
     reflect::TypeUuid,
     render::{color::Color, mesh::Mesh},
     transform::components::Transform,
@@ -89,42 +88,6 @@ impl Svg {
             },
             paths: descriptors,
             mesh: Default::default(),
-        }
-    }
-}
-
-#[derive(Clone, Component, Copy, Debug, PartialEq)]
-/// Origin of the coordinate system.
-pub enum Origin {
-    /// Bottom left of the image or viewbox.
-    BottomLeft,
-    /// Bottom right of the image or viewbox.
-    BottomRight,
-    /// Center of the image or viewbox.
-    Center,
-    /// Top left of the image or viewbox, this is the default for a SVG.
-    TopLeft,
-    /// Top right of the image or viewbox.
-    TopRight,
-}
-
-impl Default for Origin {
-    fn default() -> Self {
-        Origin::TopLeft
-    }
-}
-
-impl Origin {
-    /// Computes the translation for an origin. The resulting translation needs to be added
-    /// to the translation of the SVG.
-    pub fn compute_translation(&self, scaled_size: Vec2) -> Vec3 {
-        match self {
-            Origin::BottomLeft => Vec3::new(0.0, scaled_size.y, 0.0),
-            Origin::BottomRight => Vec3::new(-scaled_size.x, scaled_size.y, 0.0),
-            Origin::Center => Vec3::new(-scaled_size.x * 0.5, scaled_size.y * 0.5, 0.0),
-            // Standard SVG origin is top left, so we don't need to do anything
-            Origin::TopLeft => Vec3::ZERO,
-            Origin::TopRight => Vec3::new(-scaled_size.x, 0.0, 0.0),
         }
     }
 }